I hope this will fix https://github.com/fabric8-services/fabric8-auth/issues/742 :)

Side effect of this PR is slower integration test execution. On my machine the numbers the following:

With this PR:
Integration tests with no coverage: 70 sec.
Integration tests with coverage: 110 sec.

Without this PR:
Integration tests with no coverage: 20 sec.
Integration tests with coverage: 90 sec.

No changes in Unit tests. All tests are run with cleaned cache. When cache is used then of course the test run much faster. The difference in our CI/CD can be even less. It's all depends on the number of available CPUs on the host machine.